+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 
 ## 0.7a
+fixed:
+- config's warning logs
+- not saves the token if first authenticated
+
 other changes:
 - removed support of schemas.yaml