+++ CHANGELOG.md
@@ -11,6 +11,7 @@ fixed:
 - not saves the token if first authenticated for users creds
 - service not differentiate between accounts (#20) 
 - ~ path handle support of targets (#18)
+- default_secret worked wrong
 
 other changes:
 - removed support of schemas.yaml